+++ ricaun.Revit.UI.Example/Revit/App.cs
@@ -2,10 +2,8 @@ using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using Autodesk.Revit.ApplicationServices;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
-using System.Windows.Media;
+using ricaun.Revit.UI.Example.Proprieties;
 
 namespace ricaun.Revit.UI.Example.Revit
 {
@@ -19,93 +17,63 @@ namespace ricaun.Revit.UI.Example.Revit
         public Result OnStartup(UIControlledApplication application)
         {
             ribbonPanel = application.CreatePanel(TabName, PanelName);
-            var button = ribbonPanel.AddPushButton<Commands.Command>();
+
+            ribbonPanel.AddPushButton<Commands.Command>();
 
             ribbonPanel.CreatePulldownButton("PulldownButton", new[] {
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>() });
+                ribbonPanel.NewPushButtonData<Commands.Command<int>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<double>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<bool>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<string>>()
+            });
 
             ribbonPanel.CreateSplitButton("SplitButton", new[] {
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>(),
-                ribbonPanel.NewPushButtonData<Commands.Command>() });
+                ribbonPanel.NewPushButtonData<Commands.Command<int>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<double>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<bool>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<string>>()
+            });
 
-            var button2 = ribbonPanel.AddPushButton<Commands.Command>("");
+            ribbonPanel.AddPushButton<Commands.Command<Commands.Command>>();
 
-            var items = ribbonPanel.AddStackedItems(
-                ribbonPanel.NewPushButtonData<Commands.Command>("Item1"),
-                ribbonPanel.NewPushButtonData<Commands.Command>("Item2"));
+            ribbonPanel.AddStackedItems(
+                ribbonPanel.NewPushButtonData<Commands.Command<UIApplication>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<UIDocument>>());
 
-            foreach (var item in items)
-            {
-                item.SetItemSize();
-                item.SetText();
-            }
+            ribbonPanel.AddStackedItems(
+                ribbonPanel.NewPushButtonData<Commands.Command<Application>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<Document>>());
 
-            var item3s = ribbonPanel.AddStackedItems(
-                ribbonPanel.NewPushButtonData<Commands.Command>("1"),
-                ribbonPanel.NewPushButtonData<Commands.Command>("2"),
-                ribbonPanel.NewPushButtonData<Commands.Command>("3"));
 
-            foreach (var item in item3s)
-            {
-                item.SetText();
-            }
-            /*
-            foreach (var item in ribbonPanel.GetItems())
-            {
-                if (item is PushButton pushButton)
-                {
-                    pushButton.LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource();
-                    pushButton.Image = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(0.5);
-                    pushButton.ToolTipImage = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(2);
-                }
-                if (item is PulldownButton pulldownButton)
-                {
-                    pulldownButton.LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource();
-                    pulldownButton.Image = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(0.5);
-                    pulldownButton.ToolTipImage = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(2);
-                }
-                if (item is SplitButton splitButton)
-                {
-                    foreach (var i in splitButton.GetItems())
-                    {
-                        if (i is PushButton pb)
-                        {
-                            pb.LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource();
-                            pb.Image = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(0.5);
-                            pb.ToolTipImage = Proprieties.Resource.LargeImage.GetBitmapSource().Scale(2);
-                        }
-                    }
-                }
-            }
-            */
+            ribbonPanel.AddStackedItems(
+                ribbonPanel.NewPushButtonData<Commands.Command<RibbonItem>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<RibbonButton>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<RibbonPanel>>());
 
-            ribbonPanel.GetRibbonPanel().Tab.SetOrderPanels();
-            var ric = ribbonPanel.GetRibbonPanel().Tab.Panels.ToList().FirstOrDefault(e => e.Source.Title == "ricaun");
-            ric?.MoveRibbonPanel();
+            ribbonPanel.AddStackedItems(
+                ribbonPanel.NewPushButtonData<Commands.Command<Element>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<ElementType>>(),
+                ribbonPanel.NewPushButtonData<Commands.Command<ElementArray>>());
+
+            OrderPanelAndMove(ribbonPanel);
 
             foreach (var item in ribbonPanel.GetRibbonItems())
             {
                 //Console.WriteLine($"{item} {item.Name}");
             }
 
+            UpdateRibbonDescription(ribbonPanel);
+
+            return Result.Succeeded;
+        }
+
+        private void UpdateRibbonDescription(RibbonPanel ribbonPanel)
+        {
             ribbonPanel.UpdateRibbonDescription(setting =>
             {
-                setting.Add("",
+                setting.AddDefault(
                     new RibbonDescription()
                     {
-                        LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource(),
                         Help = "https://ricaun.com"
                     }
                 );
@@ -113,11 +81,10 @@ namespace ricaun.Revit.UI.Example.Revit
                 setting.Add<Commands.Command>(
                     new RibbonDescription()
                     {
+                        LargeImage = Resource.LargeImage.GetBitmapSource(),
                         Text = "Hello",
                         ToolTip = "This is a Tool Tip",
                         LongDescription = "This is a Long Description",
-                        LargeImage = Proprieties.Resource.LargeImage.GetBitmapSource(),
-                        Help = "https://ricaun.com"
                     },
                     new RibbonDescription(LanguageType.Brazilian_Portuguese)
                     {
@@ -127,26 +94,174 @@ namespace ricaun.Revit.UI.Example.Revit
                     }
                 );
 
+                setting.Add<Commands.Command<Commands.Command>>(
+                    new RibbonDescription()
+                    {
+                        Text = "Ok",
+                        LargeImage = Icons8.Ok
+                    }
+                );
+
+                setting.Add<Commands.Command<int>>(
+                    new RibbonDescription()
+                    {
+                        Text = "int",
+                        LargeImage = Icons8.Document
+                    }
+                );
+
+                setting.Add<Commands.Command<double>>(
+                    new RibbonDescription()
+                    {
+                        Text = "double",
+                        LargeImage = Icons8.File
+                    }
+                );
+
+                setting.Add<Commands.Command<bool>>(
+                    new RibbonDescription()
+                    {
+                        Text = "bool",
+                        LargeImage = Icons8.Support
+                    }
+                );
+
+                setting.Add<Commands.Command<string>>(
+                    new RibbonDescription()
+                    {
+                        Text = "text",
+                        LargeImage = Icons8.Settings
+                    }
+                );
+
                 setting.Add("PulldownButton", new RibbonDescription()
                 {
-                    Text = "PulldownButton",
-                    Help = "https://ricaun.com"
+                    Text = "Menu",
+                    LargeImage = Icons8.About
                 });
 
-                setting.Add("SplitButton", new RibbonDescription()
+
+                setting.Add<Commands.Command<UIApplication>>(new RibbonDescription()
                 {
-                    Text = "SplitButton",
-                    Help = "https://ricaun.com"
+                    Text = "UIApplication",
+                    LargeImage = Icons8.Restart,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.SetItemSize();
+                        ribbonItem.SetShowText();
+                    }
                 });
-            });
 
-            return Result.Succeeded;
+                setting.Add<Commands.Command<UIDocument>>(new RibbonDescription()
+                {
+                    Text = "UIDocument",
+                    LargeImage = Icons8.Filter,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.SetItemSize();
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<Application>>(new RibbonDescription()
+                {
+                    Text = "Application",
+                    LargeImage = Icons8.Search,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.SetItemSize();
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<Document>>(new RibbonDescription()
+                {
+                    Text = "Document",
+                    LargeImage = Icons8.Trash,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.SetItemSize();
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<RibbonItem>>(new RibbonDescription()
+                {
+                    Text = "RibbonItem",
+                    LargeImage = Icons8.Home,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.GetRibbonItem().AddQuickAccessToolBar();
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<RibbonButton>>(new RibbonDescription()
+                {
+                    Text = "RibbonButton",
+                    LargeImage = Icons8.Menu,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.GetRibbonItem().AddQuickAccessToolBar();
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<RibbonPanel>>(new RibbonDescription()
+                {
+                    Text = "RibbonPanel",
+                    LargeImage = Icons8.Info,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.GetRibbonItem().AddQuickAccessToolBar();
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<Element>>(new RibbonDescription()
+                {
+                    Text = "Element",
+                    LargeImage = Icons8.Circled,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<ElementType>>(new RibbonDescription()
+                {
+                    Text = "ElementType",
+                    LargeImage = Icons8.Checked,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+                setting.Add<Commands.Command<ElementArray>>(new RibbonDescription()
+                {
+                    Text = "ElementArray",
+                    LargeImage = Icons8.Cancel,
+                    Action = (ribbonItem) =>
+                    {
+                        ribbonItem.SetShowText();
+                    }
+                });
+
+            });
         }
 
         public Result OnShutdown(UIControlledApplication application)
         {
             ribbonPanel.Close();
             return Result.Succeeded;
+        }
+
+        private void OrderPanelAndMove(RibbonPanel ribbonPanel)
+        {
+            ribbonPanel.GetRibbonPanel().Tab.SetOrderPanels();
+            var ric = ribbonPanel.GetRibbonPanel().Tab.Panels.ToList().FirstOrDefault(e => e.Source.Title == "ricaun");
+            ric?.MoveRibbonPanel();
         }
     }
 }