+++ ricaun.Revit.UI.Example/Revit/Commands/Command.cs
@@ -12,11 +12,20 @@ namespace ricaun.Revit.UI.Example.Revit.Commands
         {
             UIApplication uiapp = commandData.Application;
 
-            //new Services.ReflectionService().GetFields<RibbonPanel>();
-            //new Services.ReflectionService().GetMethods<RibbonPanel>();
+            return Result.Succeeded;
+        }
+    }
 
-            //Console.WriteLine(typeof(Command).Assembly);
-            System.Windows.MessageBox.Show(AutodeskExtension.GetAutodeskOwner(), $"Hello Revit\n{DateTime.Now}");
+    [Transaction(TransactionMode.Manual)]
+    public class Command<T> : IExternalCommand
+    {
+        public Result Execute(ExternalCommandData commandData, ref string message, ElementSet elementSet)
+        {
+            UIApplication uiapp = commandData.Application;
+
+            var t = typeof(T);
+
+            System.Windows.MessageBox.Show(AutodeskExtension.GetAutodeskOwner(), $"Hello Revit\n{DateTime.Now}\n{t}");
 
             return Result.Succeeded;
         }