+++ ricaun.Revit.UI/RibbonPanelExtension.cs
@@ -1,4 +1,5 @@
 ﻿using Autodesk.Revit.UI;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -49,7 +50,6 @@ namespace ricaun.Revit.UI
             }
             catch
             {
-                //ribbonManager = application.GetRibbonPanels(tabName).FirstOrDefault(r => r.Name.StartsWith(panelName) && r.Visible);
                 if (ribbonManager == null)
                 {
                     ribbonManager = application.CreateRibbonPanel(tabName, SafeRibbonPanelName(panelName));
@@ -66,8 +66,12 @@ namespace ricaun.Revit.UI
         /// <returns></returns>
         public static RibbonPanel Remove(this RibbonPanel ribbonPanel)
         {
+            foreach (var ribbonItem in ribbonPanel.GetRibbonItems())
+                ribbonItem.GetRibbonItem().RemoveQuickAccessToolBar();
+
             ribbonPanel.Visible = false;
             ribbonPanel.Enabled = false;
+
             var panel = ribbonPanel.GetRibbonPanel();
             panel.Tab.Panels.Remove(panel);
             return ribbonPanel;
@@ -88,15 +92,14 @@ namespace ricaun.Revit.UI
         /// <summary>
         /// NewPushButtonData
         /// </summary>
-        /// <typeparam name="TExternalCommand"></typeparam>
         /// <param name="ribbonPanel"></param>
+        /// <param name="commandType"></param>
         /// <param name="text"></param>
         /// <returns></returns>
-        public static PushButtonData NewPushButtonData<TExternalCommand>(this RibbonPanel ribbonPanel, string text = null) where TExternalCommand : class, IExternalCommand, new()
+        public static PushButtonData NewPushButtonData(this RibbonPanel ribbonPanel, Type commandType, string text = null)
         {
-            var commandType = typeof(TExternalCommand);
-            var targetName = commandType.Name;
-            var targetText = commandType.Name;
+            var targetName = commandType.GetName();
+            var targetText = targetName;
             var assemblyName = commandType.Assembly.Location;
             var className = commandType.FullName;
 
@@ -112,6 +115,19 @@ namespace ricaun.Revit.UI
             if (text == "") currentBtn.Text = "-";
 
             return currentBtn;
+        }
+
+        /// <summary>
+        /// NewPushButtonData
+        /// </summary>
+        /// <typeparam name="TExternalCommand"></typeparam>
+        /// <param name="ribbonPanel"></param>
+        /// <param name="text"></param>
+        /// <returns></returns>
+        public static PushButtonData NewPushButtonData<TExternalCommand>(this RibbonPanel ribbonPanel, string text = null) where TExternalCommand : class, IExternalCommand, new()
+        {
+            var commandType = typeof(TExternalCommand);
+            return ribbonPanel.NewPushButtonData(commandType, text);
         }
         /// <summary>
         /// NewPushButtonData